+++ dashboard/15-final/app/dashboard/invoices/page.tsx
@@ -1,9 +1,9 @@
-import Table from "@/app/ui/table";
+import Table from "@/app/ui/invoices/table";
 
 export default function Page() {
   return (
     <div>
       <Table />
     </div>
-  )
+  );
 }
