+++ dashboard/15-final/app/dashboard/layout.tsx
@@ -1,5 +1,5 @@
-import TopNav from "@/app/ui/dashboard-topnav";
-import SideNav from "../ui/dashboard-sidenav";
+import TopNav from "@/app/ui/dashboard/topnav";
+import SideNav from "@/app/ui/dashboard/sidenav";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
